+++ packages/visual-stack-redux/src/actions.js
@@ -3,6 +3,7 @@ import { createAction, handleActions } from 'redux-actions';
 import { sortData } from '@cjdev/visual-stack/lib/components/Table/DataTable/sortingHelper';
 
 const defaultToEmpty = R.defaultTo({});
+const defaultToEmptyArray = R.defaultTo([]);
 const collapse = R.set(R.lensProp('expanded'), false);
 
 const OPEN_MODAL = '@cjdev/visual-stack-redux/OPEN_MODAL';
@@ -115,10 +116,13 @@ const findSortingIndex = (sortingOption, columns) =>
   R.findIndex(R.propEq('label', sortingOption.label))(columns);
 
 const getDataTableData = ({ columns, data, sortingOption }) => {
-  if (!sortingOption) return data;
+  const dataWithSelectedProp = defaultToEmptyArray(data).map((row, index) => {
+    return { id: index, row, selected: false };
+  });
+  if (!sortingOption) return dataWithSelectedProp;
 
   const sortingIndex = findSortingIndex(sortingOption, columns);
-  return sortData(sortingIndex, sortingOption.order, data);
+  return sortData(sortingIndex, sortingOption.order, dataWithSelectedProp);
 };
 
 const INITIALIZE_DATA_TABLE = '@cjdev/visual-stack-redux/INITIALIZE_DATA_TABLE';
@@ -132,6 +136,9 @@ const SET_DATA_TABLE_SORTING_OPTION =
 export const setDataTableSortingOption = createAction(
   SET_DATA_TABLE_SORTING_OPTION
 );
+
+const SET_DATA_TABLE_ROWS = '@cjdev/visual-stack-redux/SET_DATA_TABLE_ROWS';
+export const setDataTableRows = createAction(SET_DATA_TABLE_ROWS);
 
 export const selectDataTable = id => state => {
   const dataTable = R.view(R.lensPath(['visualStack', 'dataTable', id]))(state);
@@ -211,7 +218,6 @@ export default handleActions(
       R.set(R.lensPath(['slidingPanel', filterLabel, 'expanded']), true, state),
     [SET_SLIDING_PANEL_ACTIVE_STATE]: (state, { payload: newActiveState }) =>
       R.set(R.lensPath(['slidingPanel', 'active']), newActiveState, state),
-
     [SELECT_TAB]: (state, { payload: { tabLayoutId, index } }) => {
       return R.set(
         R.lensPath(['tabLayout', tabLayoutId, 'index']),
@@ -264,8 +270,8 @@ export default handleActions(
       ),
     [RESET_CALENDAR_SELECTION]: (state, { payload: { datePickerId } }) =>
       R.set(R.lensPath(['datePicker', datePickerId]), {}, state),
-    [INITIALIZE_DATA_TABLE]: (state, { payload }) =>
-      R.set(
+    [INITIALIZE_DATA_TABLE]: (state, { payload }) => {
+      return R.set(
         R.lensPath(['dataTable', payload.id]),
         {
           data: getDataTableData(payload),
@@ -279,24 +285,22 @@ export default handleActions(
           )(payload),
         },
         state
-      ),
+      );
+    },
     [SET_DATA_TABLE_PAGE]: (state, { payload: { id, pagination } }) =>
       R.set(R.lensPath(['dataTable', id, 'pagination']), pagination, state),
     [SET_DATA_TABLE_SORTING_OPTION]: (
       state,
-      { payload: { id, data, sortingOption } }
+      { payload: { id, sortingOption } }
     ) => {
-      const stateWithSortingOption = R.set(
+      return R.set(
         R.lensPath(['dataTable', id, 'sortingOption']),
         sortingOption,
         state
       );
-
-      return R.set(
-        R.lensPath(['dataTable', id, 'data']),
-        data,
-        stateWithSortingOption
-      );
+    },
+    [SET_DATA_TABLE_ROWS]: (state, { payload: { id, data } }) => {
+      return R.set(R.lensPath(['dataTable', id, 'data']), data, state);
     },
   },
   {
@@ -309,5 +313,6 @@ export default handleActions(
     tabLayout: {},
     pagination: {},
     datePicker: {},
+    dataTable: {},
   }
 );