+++ packages/visual-stack-redux/test/actions.test.js
@@ -17,6 +17,7 @@ import reducer, {
   initializeDataTable,
   setDataTablePage,
   setDataTableSortingOption,
+  setDataTableRows,
 } from '../src/actions';
 import { selectDataTable } from '../src/actions';
 import { DESCENDING } from '@cjdev/visual-stack/lib/components/Table/DataTable/sortingHelper';
@@ -34,6 +35,7 @@ describe('reducer', () => {
       tabLayout: {},
       pagination: {},
       datePicker: {},
+      dataTable: {},
     });
   });
 
@@ -354,11 +356,7 @@ describe('reducer', () => {
 
   test('should initialize dataTable data', () => {
     const id = 'sample-data-table';
-    const data = [
-      {
-        id: 1,
-      },
-    ];
+    const data = [[1, 'first name', 'last name']];
     const columns = [];
     const sortingOption = null;
     const beforeState = {};
@@ -366,8 +364,15 @@ describe('reducer', () => {
       beforeState,
       initializeDataTable({ id, data, columns, sortingOption })
     );
+    const dataWithSelectedProp = [
+      {
+        id: 0,
+        row: [1, 'first name', 'last name'],
+        selected: false,
+      },
+    ];
     expect(afterState.dataTable[id]).toEqual({
-      data,
+      data: dataWithSelectedProp,
       pagination: {
         page: 1,
         rowsPerPage: 10,
@@ -397,7 +402,13 @@ describe('reducer', () => {
       data: actualData,
       sortingOption: actualSortingOption,
     } = afterState.dataTable[id];
-    expect(actualData).toEqual([[7], [4], [3], [2], [1]]);
+    expect(actualData).toEqual([
+      { id: 2, row: [7], selected: false },
+      { id: 0, row: [4], selected: false },
+      { id: 1, row: [3], selected: false },
+      { id: 4, row: [2], selected: false },
+      { id: 3, row: [1], selected: false },
+    ]);
     expect(actualSortingOption).toEqual(sortingOption);
   });
 
@@ -419,7 +430,6 @@ describe('reducer', () => {
 
   test('should setDataTableSortingOption', () => {
     const id = 'sample-data-table';
-    const data = [{ id: 1 }];
     const sortingOption = {
       label: 'id',
       order: 'ASCENDING',
@@ -427,11 +437,20 @@ describe('reducer', () => {
     const beforeState = {};
     const afterState = reducer(
       beforeState,
-      setDataTableSortingOption({ id, data, sortingOption })
+      setDataTableSortingOption({ id, sortingOption })
     );
     expect(afterState.dataTable[id]).toEqual({
-      data,
       sortingOption,
+    });
+  });
+
+  test('should setDataTableRows', () => {
+    const id = 'sample-data-table';
+    const data = [{ id: 0, row: [1], selected: false }];
+    const beforeState = {};
+    const afterState = reducer(beforeState, setDataTableRows({ id, data }));
+    expect(afterState.dataTable[id]).toEqual({
+      data,
     });
   });