+++ packages/visual-stack/src/components/Table/DataTable/index.js
@@ -17,10 +17,17 @@ import PropTypes from 'prop-types';
 import './DataTable.css';
 import LoadingAnimation from '../../LoadingAnimation';
 
-const generateHeader = ({ sortable, sortingOption, onSort, data }) => (
-  { label: currentLabel, width },
-  index
+const generateHeaders = ({ columns, ...tableProps }) => {
+  return columns.map((column, index) => generateHeader(column, index, tableProps));
+}
+
+const generateHeader = (
+  column,
+  index,
+  tableProps,
 ) => {
+  const { sortable, sortingOption, onSort, data } = tableProps;
+  const { label: currentLabel, width } = column;
   const isCurrentColumnSorted = () => {
     return sortingOption.label === currentLabel;
   };
@@ -58,9 +65,21 @@ const generateHeader = ({ sortable, sortingOption, onSort, data }) => (
   );
 };
 
-const generateRow = ({ onClick, columns }) => (rowItems, index) => (
-  <Tr key={index}>
-    {rowItems.map((item, columnIndex) => {
+const generateRows = ({ data, ...tableProps }) => {
+  return data.map((rowData, index) => generateRow(rowData, index, tableProps));
+}
+
+const generateRow = (rowData, index, tableProps) => {
+  const { onClick, columns, selectable, onSelect } = tableProps;
+  const { id, row, selected } = rowData;
+
+  return <Tr key={id}>
+    {selectable && <Td>
+      <input type="checkbox" aria-label="Select row from data table" checked={selected} onChange={() => {
+        onSelect(id)
+      }} />
+    </Td>}
+    {row.map((item, columnIndex) => {
       const getColumn = R.compose(
         R.defaultTo({}),
         R.prop(columnIndex)
@@ -87,7 +106,7 @@ const generateRow = ({ onClick, columns }) => (rowItems, index) => (
       );
     })}
   </Tr>
-);
+};
 
 const getDataWithPagination = (rowsPerPage, page) =>
   R.compose(
@@ -98,6 +117,23 @@ const getDataWithPagination = (rowsPerPage, page) =>
 const NoDataLabel = ({ label }) => {
   return <div className="vs-data-table-no-data-label">{label}</div>;
 };
+
+const omitGeneratedIdsFromData = data => R.map(R.dissoc("id"))(data)
+
+
+const toggleSelectedInRowData = rowData => ({
+  ...rowData,
+  selected: !rowData.selected
+})
+
+const selectAllCheckbox = ({ data, onSelect }) => {
+  const checked = R.all(R.propEq('selected', true))(data);
+  let onChange = () => onSelect({
+    data: R.map(rowData => ({ ...rowData, selected: !checked }))(data)
+  });
+  return <input type="checkbox" aria-label="Select all from data table" checked={checked} onChange={(onChange)} />
+}
+
 
 export const DataTable = ({
   caption = '',
@@ -118,11 +154,19 @@ export const DataTable = ({
   noDataLabel = 'No data available.',
   rowsPerPageTemplate,
   totalRecordsTemplate,
+  selectable = false,
+  onSelect,
   ...restProps
 }) => {
-  const normalizedData = pagination
+  const paginatedData = pagination
     ? getDataWithPagination(rowsPerPage, page)(data)
     : data;
+
+  const onSelectId = id => {
+    const selectedRowIndex = R.findIndex(R.propEq('id', id), data);
+    const dataWithSelectedRow = R.adjust(selectedRowIndex, toggleSelectedInRowData)(data);
+    onSelect({ data: dataWithSelectedRow })
+  }
 
   return (
     <TableContainer {...restProps} className="vs-data-table-container">
@@ -132,7 +176,7 @@ export const DataTable = ({
             {caption}
             <p>{description}</p>
           </div>
-          <div>{renderToolbar && renderToolbar({ columns, data })}</div>
+          <div>{renderToolbar && renderToolbar({ columns, data: omitGeneratedIdsFromData(data) })}</div>
         </div>
       </TableTitle>
       {isLoading ? (
@@ -140,33 +184,46 @@ export const DataTable = ({
           <LoadingAnimation loadingMessage={loadingMessage} />
         </div>
       ) : (
-        <>
-          <Table>
-            <THead>
-              <Tr>
-                {columns.map(
-                  generateHeader({ sortable, sortingOption, onSort, data })
-                )}
-              </Tr>
-            </THead>
-            <TBody>
-              {normalizedData.map(generateRow({ onClick, columns }))}
-            </TBody>
-          </Table>
-          {normalizedData.length === 0 && <NoDataLabel label={noDataLabel} />}
-          {pagination && (
-            <Pagination
-              className="vs-table-pagination"
-              numberOfRows={data.length}
-              rowsPerPage={rowsPerPage}
-              page={page}
-              onChange={onPageChange}
-              rowsPerPageTemplate={rowsPerPageTemplate}
-              totalRecordsTemplate={totalRecordsTemplate}
-            />
-          )}
-        </>
-      )}
+          <>
+            <Table>
+              <THead>
+                <Tr>
+                  {selectable && <Th>
+                    {selectAllCheckbox({ data, onSelect })}
+                  </Th>}
+                  {generateHeaders({
+                    columns,
+                    sortable,
+                    sortingOption,
+                    onSort,
+                    data
+                  })}
+                </Tr>
+              </THead>
+              <TBody>
+                {generateRows({
+                  data: paginatedData,
+                  onClick,
+                  columns,
+                  selectable,
+                  onSelect: onSelectId
+                })}
+              </TBody>
+            </Table>
+            {paginatedData.length === 0 && <NoDataLabel label={noDataLabel} />}
+            {pagination && (
+              <Pagination
+                className="vs-table-pagination"
+                numberOfRows={data.length}
+                rowsPerPage={rowsPerPage}
+                page={page}
+                onChange={onPageChange}
+                rowsPerPageTemplate={rowsPerPageTemplate}
+                totalRecordsTemplate={totalRecordsTemplate}
+              />
+            )}
+          </>
+        )}
     </TableContainer>
   );
 };
@@ -188,6 +245,8 @@ DataTable.propTypes = {
   }),
   sortable: PropTypes.bool,
   onSort: PropTypes.func,
+  selectable: PropTypes.bool,
+  onSelect: PropTypes.func,
   onClick: PropTypes.func,
   renderToolbar: PropTypes.func,
   noDataLabel: PropTypes.string,