+++ packages/visual-stack/src/components/Table/DataTable/tests/index.test.js
@@ -12,7 +12,11 @@ describe('DataTable', () => {
   describe('Top Level Header', () => {
     it('should display tool bar', () => {
       const renderToolbar = jest.fn(() => <div id="test">TEST</div>);
-      const data = [[1, 2, 3]];
+      const data = [{ id: 0, row: [1, 2, 3], selected: false }, { id: 1, row: [4, 5, 6], selected: true }];
+      const dataWithoutIds = [
+        { row: data[0].row, selected: data[0].selected },
+        { row: data[1].row, selected: data[1].selected },
+      ]
       const columns = [
         {
           label: 'hello',
@@ -25,7 +29,7 @@ describe('DataTable', () => {
           columns={columns}
         />
       );
-      expect(renderToolbar.mock.calls).toEqual([[{ data, columns }]]);
+      expect(renderToolbar.mock.calls).toEqual([[{ data: dataWithoutIds, columns }]]);
       expect(wrapper.find('div#test').text()).toEqual('TEST');
     });
   });
@@ -100,7 +104,10 @@ describe('DataTable', () => {
       const data3 = 'data3';
       const data4 = 'data4';
       const wrapper = mount(
-        <DataTable data={[[data1, data2], [data3, data4]]} />
+        <DataTable data={[
+          { id: 0, row: [data1, data2], selected: false },
+          { id: 1, row: [data3, data4], selected: false }]
+        } />
       );
       expect(wrapper.find('td').map(node => node.text())).toEqual([
         data1,
@@ -134,10 +141,10 @@ describe('DataTable', () => {
         <DataTable
           columns={columns}
           data={[
-            [1, 2, 3, 4],
-            [1, 2, 3, 4],
-            [1, 2, targetData, 4],
-            [1, 2, 3, 4],
+            { id: 0, row: [1, 2, 3, 4], selected: false },
+            { id: 1, row: [1, 2, 3, 4], selected: false },
+            { id: 2, row: [1, 2, targetData, 4], selected: false },
+            { id: 3, row: [1, 2, 3, 4], selected: false },
           ]}
           onClick={onClick}
         />
@@ -167,7 +174,7 @@ describe('DataTable', () => {
       const wrapper = mount(
         <DataTable
           columns={columns}
-          data={[[targetData, 2], [1, 2]]}
+          data={[{ id: 0, row: [targetData, 2], selected: false }, { id: 1, row: [1, 2], selected: false }]}
           onClick={onClick}
         />
       );
@@ -178,7 +185,7 @@ describe('DataTable', () => {
       expect(onClick.mock.calls).toHaveLength(0);
     });
 
-    describe('Displaying custom cell', () => {
+    describe('displaying custom cell', () => {
       it('should display custom cell when there is renderCell function', () => {
         const label = 'change';
         const targetData = 'value';
@@ -190,7 +197,7 @@ describe('DataTable', () => {
           },
         ];
         const wrapper = mount(
-          <DataTable columns={columns} data={[[targetData]]} />
+          <DataTable columns={columns} data={[{ id: 0, row: [targetData], selected: false }]} />
         );
         const targetTD = wrapper
           .find('td.vs-cell')
@@ -198,6 +205,124 @@ describe('DataTable', () => {
         expect(targetTD).toHaveLength(1);
       });
     });
+
+    describe("selecting rows", () => {
+      it("should not display any selectable rows if table is not selectable", () => {
+        const data = [{ id: 0, row: [1], selected: true, }, { id: 1, row: [2], selected: false, }]
+        const wrapper = mount(
+          <DataTable data={data} selectable={false} />
+        );
+        const rowCheckboxes = wrapper.find("input[aria-label='Select row from data table']");
+        const selectAllCheckbox = wrapper.find("input[aria-label='Select all from data table']");
+        expect(rowCheckboxes).toHaveLength(0);
+        expect(selectAllCheckbox).toHaveLength(0);
+      })
+
+      it("should display selected rows", () => {
+        const data = [
+          { id: 0, row: [1], selected: true, },
+          { id: 1, row: [2], selected: false, },
+        ]
+        const wrapper = mount(
+          <DataTable data={data} selectable />
+        );
+        const firstRowSelected = wrapper.find("input[aria-label='Select row from data table']").at(0).prop("checked")
+        const secondRowSelected = wrapper.find("input[aria-label='Select row from data table']").at(1).prop("checked")
+
+        expect(firstRowSelected).toEqual(true);
+        expect(secondRowSelected).toEqual(false);
+      })
+
+      it("should select one row", () => {
+        const onSelect = jest.fn();
+        const data = [{ id: 0, row: [1], selected: false, }, { id: 1, row: [2], selected: false, }]
+        const wrapper = mount(
+          <DataTable data={data} selectable onSelect={onSelect} />
+        );
+        wrapper.find("input[aria-label='Select row from data table']").at(0).simulate('change');
+
+        expect(onSelect).lastCalledWith({
+          data: [{
+            id: 0, row: [1], selected: true
+          },
+          {
+            id: 1, row: [2], selected: false
+          }]
+        })
+      })
+
+      it("should check selected all checkbox when all rows are selected", () => {
+        const data = [
+          { id: 0, row: [1], selected: true, },
+          { id: 1, row: [2], selected: true, },
+        ]
+        const wrapper = mount(
+          <DataTable data={data} selectable />
+        );
+
+        const selectedAllRowsCheckbox = wrapper.find("input[aria-label='Select all from data table']").prop("checked")
+
+        expect(selectedAllRowsCheckbox).toEqual(true);
+      })
+
+      it("should uncheck selected all checkbox when some rows are not selected", () => {
+        const data = [
+          { id: 0, row: [1], selected: true, },
+          { id: 1, row: [2], selected: false, },
+        ]
+        const wrapper = mount(
+          <DataTable data={data} selectable />
+        );
+
+        const selectedAllRowsCheckbox = wrapper.find("input[aria-label='Select all from data table']").prop("checked")
+
+        expect(selectedAllRowsCheckbox).toEqual(false);
+      })
+
+      it("should select all rows at once when clicking on select all checkbox", () => {
+        const data = [
+          { id: 0, row: [1], selected: false, },
+          { id: 1, row: [2], selected: false, },
+        ]
+        const onSelect = jest.fn();
+
+        const wrapper = mount(
+          <DataTable data={data} selectable onSelect={onSelect} />
+        );
+
+        const selectedAllRowsCheckbox = wrapper.find("input[aria-label='Select all from data table']");
+        selectedAllRowsCheckbox.simulate('change');
+
+        expect(onSelect).lastCalledWith({
+          data: [
+            { id: 0, row: [1], selected: true, },
+            { id: 1, row: [2], selected: true, },
+          ]
+        });
+      })
+
+      it("should unselect all rows at once when clicking on select all checkbox", () => {
+        const data = [
+          { id: 0, row: [1], selected: true, },
+          { id: 1, row: [2], selected: true, },
+        ]
+        const onSelect = jest.fn();
+
+        const wrapper = mount(
+          <DataTable data={data} selectable onSelect={onSelect} />
+        );
+
+        const selectedAllRowsCheckbox = wrapper.find("input[aria-label='Select all from data table']");
+        selectedAllRowsCheckbox.simulate('change');
+
+        expect(onSelect).lastCalledWith({
+          data: [
+            { id: 0, row: [1], selected: false, },
+            { id: 1, row: [2], selected: false, },
+          ]
+        });
+      })
+    })
   });
 
   describe('pagination', () => {
@@ -255,9 +380,9 @@ describe('DataTable', () => {
       const group1 = range(1, 26);
       const group2 = range(26, 51);
       const group3 = range(51, 76);
-      const groupedData1 = group1.map(number => [number]);
-      const groupedData2 = group2.map(number => [number]);
-      const groupedData3 = group3.map(number => [number]);
+      const groupedData1 = group1.map(number => ({ id: number, row: [number], selected: false }));
+      const groupedData2 = group2.map(number => ({ id: number, row: [number], selected: false }));
+      const groupedData3 = group3.map(number => ({ id: number, row: [number], selected: false }));
 
       it('should render a fixed number of rows when pagination is enabled', () => {
         const onPageChange = jest.fn();
@@ -513,7 +638,11 @@ describe('DataTable', () => {
               label: 'NOT YOUR TARGET HEADER',
               order: DESCENDING,
             }}
-            data={[[second], [first], [third]]}
+            data={[
+              { id: 0, row: [second], selected: false },
+              { id: 1, row: [first], selected: false },
+              { id: 2, row: [third], selected: false }
+            ]}
             onSort={onSort}
             sortable
           />
@@ -524,9 +653,9 @@ describe('DataTable', () => {
         targetHeaderWrapper.simulate('click');
 
         expect(onSort.mock.calls[0][0].data).toEqual([
-          [third],
-          [second],
-          [first],
+          { id: 2, row: [third], selected: false },
+          { id: 0, row: [second], selected: false },
+          { id: 1, row: [first], selected: false },
         ]);
       });
 
@@ -547,7 +676,11 @@ describe('DataTable', () => {
               label,
               order: DESCENDING,
             }}
-            data={[[second], [first], [third]]}
+            data={[
+              { id: 0, row: [second], selected: false },
+              { id: 1, row: [first], selected: false },
+              { id: 2, row: [third], selected: false }
+            ]}
             onSort={onSort}
             sortable
           />
@@ -558,9 +691,9 @@ describe('DataTable', () => {
         targetHeaderWrapper.simulate('click');
 
         expect(onSort.mock.calls[0][0].data).toEqual([
-          [first],
-          [second],
-          [third],
+          { id: 1, row: [first], selected: false },
+          { id: 0, row: [second], selected: false },
+          { id: 2, row: [third], selected: false },
         ]);
       });
 
@@ -581,7 +714,10 @@ describe('DataTable', () => {
               label,
               order: ASCENDING,
             }}
-            data={[[second], [first], [third]]}
+            data={[
+              { id: 0, row: [second], selected: false },
+              { id: 1, row: [first], selected: false },
+              { id: 2, row: [third], selected: false }]}
             onSort={onSort}
             sortable
           />
@@ -591,9 +727,9 @@ describe('DataTable', () => {
           .filterWhere(node => trim(node.text()) === label);
         targetHeaderWrapper.simulate('click');
         expect(onSort.mock.calls[0][0].data).toEqual([
-          [third],
-          [second],
-          [first],
+          { id: 2, row: [third], selected: false },
+          { id: 0, row: [second], selected: false },
+          { id: 1, row: [first], selected: false },
         ]);
       });
     });