+++ packages/visual-stack/src/index.js
@@ -3,6 +3,7 @@ import './global';
 // modules that do not need qualification. i.e., the components
 // they export can be scoped to VisualStack.<component>
 export * from './components/Button';
+export * from './components/Badge';
 export * from './components/Spinner';
 
 // modules that need scoping. i.e. the components they