+++ CHANGELOG.md
@@ -1,3 +1,36 @@
+# 7.1.0
+
+## Updates
+
+- Add new Badge variants
+
+# 7.0.2
+
+## Fix
+
+- Fix Badge PropTypes
+
+# 7.0.1
+
+## Updates
+
+- Update Badge docs
+
+# 7.0.0
+
+## New Feature
+
+- Add selectable rows to DataTable.
+
+## Breaking Changes in DataTable
+
+- renderToolbar's data prop now return an array of objects instead of array of arrays.
+- Before: data = [['first', 'last']], After: data = [{row: ['first', 'last'], selected: true}]
+
+## New Component
+
+- Add Badge component
+
 # 6.10.0
 
 ## New Component